refactor: 라우터 별로 코드 정리하기
